pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "igniteae"
description = "IgniteAE."
readme = {file = "README.md", content-type="text/markdown"}
license = {file = "LICENSE"}
authors = [
    {name = "Taras Savchyn", email="trsvchn@gmail.com"},
]
requires-python = ">=3.11"
dependencies = [
    "torch",
    "pytorch-ignite",
    "igniteae",
]
dynamic = ["version"]

[tool.setuptools]
zip-safe = false
include-package-data = false

[tool.setuptools.packages.find]
include = ["igniteae", "igniteae.*"]

[tool.setuptools.dynamic.version]
attr = "igniteae.__version__"
